[Jax][Pallas][Mosaic] Implement platform dependent diag, with branch selection driven by constant prop in mosaic lowering.
This CL builds out a simple sketch of constant prop by construction in mosaic - we create a subclass for Value that allows us to keep the original value around, and use it to remove min and max (for now, can use anywhere!) as well as to do specific branch selection in `cond`. This specific branch selection mechanism actually allows us to go from trying to lower all the platform_dependent branches in mosaic, to lowering the winning one.

This then allows us to pick a single branch, and ignore branches which do not have a viable mosaic implementation.

And, finally, for diag, this means we can replace the initial gather-dependent implementation in lax with a mosaic specific one that avoids gather.

PiperOrigin-RevId: 693594640